stablecoin-analytics-platform/requirements.txt
web3==6.11.3
solana==0.30.2
solders==0.18.1
base58==2.1.1
psycopg[binary]==3.2.12
requests==2.31.0
pandas==2.2.3
python-dotenv==1.0.0